Branch disps are now inserted into the Disp index.
PEEK-287

+++ peek_plugin_diagram/_private/worker/tasks/ImportDispTask.py
@@ -1,10 +1,11 @@
-import ujson as json
 import logging
-import pytz
 from datetime import datetime
+from typing import List, Dict, Tuple, Optional
+
+import pytz
+import ujson as json
 from geoalchemy2.shape import to_shape
 from txcelery.defer import DeferrableTask
-from typing import List, Dict, Tuple
 from vortex.Payload import Payload
 
 from peek_plugin_base.worker import CeleryDbConn
@@ -84,7 +85,7 @@ def importDispsTask(self, modelSetKey: str, coordSetKey: str,
             coordSet, disps
         )
 
-        _bulkLoadDispsTask(coordSet, importGroupHash, ormDisps)
+        _bulkLoadDispsTask(coordSet, ormDisps, importGroupHash)
 
         liveDbImportTuples = importDispLinks(
             coordSet, importGroupHash, dispLinkImportTuples
@@ -317,7 +318,9 @@ def _convertImportTuple(importDisp):
     return disp
 
 
-def _bulkLoadDispsTask(coordSet: ModelCoordSet, importGroupHash: str, disps: List):
+def _bulkLoadDispsTask(coordSet: ModelCoordSet, disps: List,
+                       importGroupHash: Optional[str] = None,
+                       ormSession=None):
     """ Import Disps Links
 
     1) Drop all disps with matching importGroupHash
@@ -334,11 +337,15 @@ def _bulkLoadDispsTask(coordSet: ModelCoordSet, importGroupHash: str, disps: Lis
 
     dispTable = DispBase.__table__
 
-    ormSession = CeleryDbConn.getDbSession()
+    closeOrmSession = bool(ormSession)
+    if closeOrmSession:
+        ormSession = CeleryDbConn.getDbSession()
+
     try:
-        ormSession.execute(dispTable
-                           .delete()
-                           .where(dispTable.c.importGroupHash == importGroupHash))
+        if importGroupHash:
+            ormSession.execute(dispTable
+                               .delete()
+                               .where(dispTable.c.importGroupHash == importGroupHash))
 
         # Initialise the ModelCoordSet initial position if it's not set
         if (not coordSet.initialPanX
@@ -364,4 +371,5 @@ def _bulkLoadDispsTask(coordSet: ModelCoordSet, importGroupHash: str, disps: Lis
                     len(disps), (datetime.now(pytz.utc) - startTime))
 
     finally:
-        ormSession.close()
+        if closeOrmSession:
+            ormSession.close()
